Move generix maxsum def in maxsum.py
generic function and variable should be defined in maxsum.py and reused
in amaxsum.py
