+++ pydcop/algorithms/maxsum.py
@@ -38,20 +38,23 @@ Synchronous implementation of the MaxSum algorithm
 
 """
 import logging
-from typing import Optional, List, Dict, Any
+from typing import Optional, List, Dict, Any, Tuple, Union
 from collections import defaultdict
 
 
 from pydcop.algorithms import ComputationDef, AlgoParameterDef
-from pydcop.algorithms import amaxsum
-from pydcop.algorithms.amaxsum import MaxSumMessage
-from pydcop.dcop.objects import Variable
+from pydcop.computations_graph.factor_graph import (
+    FactorComputationNode,
+    VariableComputationNode,
+)
+from pydcop.dcop.objects import Variable, VariableNoisyCostFunc
 from pydcop.dcop.relations import Constraint, generate_assignment_as_dict
 from pydcop.infrastructure.computations import (
     DcopComputation,
     SynchronousComputationMixin,
     VariableComputation,
     register,
+    Message,
 )
 
 GRAPH_TYPE = "factor_graph"
@@ -61,6 +64,13 @@ logger = logging.getLogger("pydcop.maxsum")
 # (de)serialized
 # INFINITY = float('inf')
 INFINITY = 100000
+
+HEADER_SIZE = 0
+UNIT_SIZE = 1
+
+# constants for memory costs and capacity
+FACTOR_UNIT_SIZE = 1
+VARIABLE_UNIT_SIZE = 1
 
 
 def build_computation(comp_def: ComputationDef):
@@ -72,10 +82,139 @@ def build_computation(comp_def: ComputationDef):
         return FactorComputation(comp_def=comp_def)
 
 
-# MaxSum and AMaxSum have the same definitions for communication load
-# and computation footprints.
-computation_memory = amaxsum.computation_memory
-communication_load = amaxsum.communication_load
+def computation_memory(
+    computation: Union[FactorComputationNode, VariableComputationNode]
+) -> float:
+    """Memory footprint associated with the maxsum computation node.
+
+    Notes
+    -----
+    Two formulations of the memory footprint are possible for factors :
+    * If the constraint is given by a function (intentional), the factor
+      only needs to keep the costs sent by each variable and the footprint
+      is the total size of these cost vectors.
+    * If the constraints is given extensively the size of the hypercube of
+      costs must also be accounted for.
+
+    Parameters
+    ----------
+    computation: FactorComputationNode or VariableComputationNode
+        A computation node for a factor or a variable in the factor-graph.
+
+    Returns
+    -------
+    float:
+        the memory footprint of the computation.
+    """
+    if isinstance(computation, FactorComputationNode):
+        # Memory footprint associated with the factor computation f.
+        # For Maxsum, it depends on the size of the domain of the neighbor
+        # variables.
+        m = 0
+        for v in computation.variables:
+            domain_size = len(v.domain)
+            m += domain_size * FACTOR_UNIT_SIZE
+        return m
+
+    elif isinstance(computation, VariableComputationNode):
+        # For Maxsum, the memory footprint a variable computations depends
+        #  on the number of  neighbors in the factor graph.
+        domain_size = len(computation.variable.domain)
+        num_neighbors = len(list(computation.links))
+        return num_neighbors * domain_size * VARIABLE_UNIT_SIZE
+
+    raise ValueError(
+        "Invalid computation node type {}, maxsum only defines "
+        "VariableComputationNodeand FactorComputationNode".format(computation)
+    )
+
+
+def communication_load(
+    src: Union[FactorComputationNode, VariableComputationNode], target: str
+) -> float:
+    """The communication cost of an edge between a variable and a factor.
+
+    Parameters
+    ----------
+    src: VariableComputationNode
+        The ComputationNode for the source variable.
+    target: str
+        the name of the other variable `src` is sending messages to
+
+    Return
+    ------
+    float:
+        the size of messages between computation and target.
+    """
+    if isinstance(src, VariableComputationNode):
+        d_size = len(src.variable.domain)
+        return UNIT_SIZE * d_size + HEADER_SIZE
+
+    elif isinstance(src, FactorComputationNode):
+        for v in src.variables:
+            if v.name == target:
+                d_size = len(v.domain)
+                return UNIT_SIZE * d_size + HEADER_SIZE
+        raise ValueError(
+            "Could not find variable {} in constraint of factor "
+            "{}".format(target, src)
+        )
+
+    raise ValueError(
+        "maxsum communication_load only supports "
+        "VariableComputationNode and FactorComputationNode, "
+        "invalid computation: " + str(src)
+    )
+
+
+class MaxSumMessage(Message):
+    def __init__(self, costs: Dict):
+        super().__init__("max_sum", None)
+        self._costs = costs
+
+    @property
+    def costs(self):
+        return self._costs
+
+    @property
+    def size(self):
+        # Max sum messages are dictionaries from values to costs:
+        return len(self._costs) * 2
+
+    def __str__(self):
+        return "MaxSumMessage({})".format(self._costs)
+
+    def __repr__(self):
+        return "MaxSumMessage({})".format(self._costs)
+
+    def __eq__(self, other):
+        if type(other) != MaxSumMessage:
+            return False
+        if self.costs == other.costs:
+            return True
+        return False
+
+    def _simple_repr(self):
+        r = {"__module__": self.__module__, "__qualname__": self.__class__.__qualname__}
+
+        # When building the simple repr when transform the dict into a pair
+        # of list to avoid problem when serializing / de-serializing the repr.
+        # The costs dic often contains int as key, when converting to an from
+        # json (which only support string for keys in dict), we would
+        # otherwise loose the type information and restore the dict with str
+        # keys.
+        vals, costs = zip(*self._costs.items())
+        r["vals"] = vals
+        r["costs"] = costs
+        return r
+
+    @classmethod
+    def _from_repr(cls, r):
+        vals = r["vals"]
+        costs = r["costs"]
+
+        return MaxSumMessage(dict(zip(vals, costs)))
+
 
 # Some semantic type definition, to make things easier to read and check:
 VarName = str
